requirements.txt
python-dotenv
google.generativeai
pandas
streamlit